requirements.txt
Flask
flask_sqlalchemy
flask_migrate
flask_login
plotly
flask_wtf